Extract windows check to separate class
